Massive restructuration of the project layout for easier maintenance.
This is the layout used for all the third-party apps I maintain.
