+++ .gitignore
@@ -1,8 +1,7 @@
-*.pyc
-*.pyo
-/build
-/dist
-/temp
-/src/django_picklefield.egg-info
-.tox
+syntax:glob
+*.py[co]
+dist/
+django_picklefield.egg-info/*
+build/
 .coverage
+.tox
