+++ .travis.yml
@@ -1,42 +1,54 @@
+dist: trusty
 sudo: false
 language: python
+env:
+- TOXENV=flake8
+- TOXENV=isort
+cache: pip
 matrix:
-  include:
-    - env: TOXENV=flake8
-    - env: TOXENV=isort
-    - env: TOXENV=py27-1.11
-    - env: TOXENV=py34-1.11
-      python: 3.4
-    - env: TOXENV=py34-2.0
-      python: 3.4
-    - env: TOXENV=py35-1.11
-      python: 3.5
-    - env: TOXENV=py35-2.0
-      python: 3.5
-    - env: TOXENV=py35-2.1
-      python: 3.5
-    - env: TOXENV=py35-master
-      python: 3.5
-    - env: TOXENV=py36-1.11
-      python: 3.6
-    - env: TOXENV=py36-2.0
-      python: 3.6
-    - env: TOXENV=py36-2.1
-      python: 3.6
-    - env: TOXENV=py36-master
-      python: 3.6
   fast_finish: true
+  include:
+  - python: 2.7
+    env: TOXENV=py27-1.11
+  - python: 3.4
+    env: TOXENV=py34-1.11
+  - python: 3.4
+    env: TOXENV=py34-2.0
+  - python: 3.5
+    env: TOXENV=py35-1.11
+  - python: 3.5
+    env: TOXENV=py35-2.0
+  - python: 3.5
+    env: TOXENV=py35-2.1
+  - python: 3.5
+    env: TOXENV=py35-master
+  - python: 3.6
+    env: TOXENV=py36-1.11
+  - python: 3.6
+    env: TOXENV=py36-2.0
+  - python: 3.6
+    env: TOXENV=py36-2.1
+  - python: 3.6
+    env: TOXENV=py36-master
+  - python: 3.7
+    env: TOXENV=py37-2.0
+    dist: xenial
+    sudo: true
+  - python: 3.7
+    env: TOXENV=py37-2.1
+    dist: xenial
+    sudo: true
+  - python: 3.7
+    env: TOXENV=py37-master
+    dist: xenial
+    sudo: true
   allow_failures:
-    - env: TOXENV=py35-master
-      python: 3.5
-    - env: TOXENV=py36-master
-      python: 3.6
-
+  - env: TOXENV=py35-master
+  - env: TOXENV=py36-master
+  - env: TOXENV=py37-master
 install:
-  - pip install tox coveralls
-
+- pip install tox coveralls
 script:
-  - tox
-
+- tox
 after_success:
-  - coveralls
+- if [ -f .coverage ]; then coveralls; fi
