+++ picklefield/__init__.py
@@ -1,0 +1,6 @@
+from __future__ import unicode_literals
+
+from picklefield.constants import DEFAULT_PROTOCOL
+from picklefield.fields import PickledObjectField
+
+__all__ = 'DEFAULT_PROTOCOL', 'PickledObjectField'
