+++ picklefield/constants.py
@@ -1,0 +1,3 @@
+from __future__ import unicode_literals
+
+DEFAULT_PROTOCOL = 2
