--- src/picklefield/__init__.py
@@ -1,5 +1,0 @@
-"""Pickle field implementation for Django."""
-
-DEFAULT_PROTOCOL = 2
-
-from picklefield.fields import PickledObjectField  # noqa
