+++ tests/settings.py
@@ -1,13 +1,13 @@
+from __future__ import unicode_literals
+
+SECRET_KEY = 'not-anymore'
+
 DATABASES = {
     'default': {
         'ENGINE': 'django.db.backends.sqlite3',
-    },
+    }
 }
 
 INSTALLED_APPS = [
-    'picklefield',
+    'tests',
 ]
-
-SECRET_KEY = 'local'
-
-SILENCED_SYSTEM_CHECKS = ['1_7.W001']
