+++ tox.ini
@@ -1,42 +1,41 @@
 [tox]
+skipsdist = true
+args_are_paths = false
 envlist =
     flake8,
+    isort,
     py27-1.11,
-    py34-{1.11,2.0}
-    py35-{1.11,2.0,2.1,master}
-    py36-{1.11,2.0,2.1,master}
+    py34-{1.11,2.0},
+    py35-{1.11,2.0,2.1,master},
+    py36-{1.11,2.0,2.1,master},
+    py37-{1.11,2.0,2.1,master}
 
 [testenv]
 basepython =
     py27: python2.7
-    py33: python3.3
     py34: python3.4
     py35: python3.5
     py36: python3.6
+    py37: python3.7
 usedevelop = true
 commands =
-    {envpython} -R -Wonce {envbindir}/coverage run {envbindir}/django-admin.py test  --pythonpath=. --settings test_settings picklefield
+    {envpython} -R -Wonce {envbindir}/coverage run -m django test -v2 --settings=tests.settings {posargs}
     coverage report
 deps =
     coverage
-    1.8: Django>=1.8,<1.9
-    1.9: Django>=1.9,<1.10
-    1.10: Django>=1.10,<1.11
     1.11: Django>=1.11,<2.0
     2.0: Django>=2.0,<2.1
-    2.1: Django>=2.1a1,<2.2
+    2.1: Django>=2.1,<2.2
     master: https://github.com/django/django/archive/master.tar.gz
 
 [testenv:flake8]
+usedevelop = false
 basepython = python2.7
-commands =
-    flake8
-deps =
-    flake8==3.3.0
+commands = flake8
+deps = flake8
 
 [testenv:isort]
+usedevelop = false
 basepython = python2.7
-commands =
-    isort --recursive --check-only --diff src/picklefield
-deps =
-    isort==4.2.5
+commands = isort --recursive --check-only --diff picklefield tests
+deps = isort==4.2.5
